requirements.txt
flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.1
openai>=1.3.7
psycopg2-binary
chromadb